requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
opencv-python-headless
opencv-contrib-python-headless
numpy
Pillow
google-cloud-vision
